440-connector-nextgen/src/main/java/io/harness/connector/mappers/sumologicmapper/SumoLogicDTOToEntity.java
package io.harness.connector.mappers.sumologicmapper;

import static io.harness.annotations.dev.HarnessTeam.CV;

import io.harness.annotations.dev.OwnedBy;
import io.harness.connector.entities.embedded.sumologic.SumoLogicConnector;
import io.harness.connector.mappers.ConnectorDTOToEntityMapper;
import io.harness.delegate.beans.connector.sumologic.SumoLogicConnectorDTO;
import io.harness.encryption.SecretRefHelper;

@OwnedBy(CV)
public class SumoLogicDTOToEntity implements ConnectorDTOToEntityMapper<SumoLogicConnectorDTO, SumoLogicConnector> {
  @Override
  public SumoLogicConnector toConnectorEntity(SumoLogicConnectorDTO connectorDTO) {
    return SumoLogicConnector.builder()
        .url(connectorDTO.getUrl())
        .accessIdRef(SecretRefHelper.getSecretConfigString(connectorDTO.getAccessIdRef()))
        .accessKeyRef(SecretRefHelper.getSecretConfigString(connectorDTO.getAccessKeyRef()))
        .build();
  }
}
